refactor: firebase CRUD 로직 서비스 파일로 분리
